requirements.txt
accelerate==0.27.2
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apache-airflow==2.8.1
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.0
apache-beam==2.52.0
apispec==6.4.0
appdirs==1.4.4
appnope==0.1.3
argcomplete==3.2.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
clickclick==20.10.2
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
comm==0.2.1
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
contextily==1.5.0
contourpy==1.2.0
crcmod==1.7
cron-descriptor==1.4.3
croniter==2.0.1
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
cykhash==2.0.1
Cython==3.0.8
databricks-cli==0.18.0
datasets==2.17.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.4.2
docker==7.0.0
docopt==0.6.2
docutils==0.20.1
email-validator==2.1.0.post1
entrypoints==0.4
evidently==0.4.14
executing==2.0.1
fastapi==0.109.0
fastapi-restful==0.5.0
fastavro==1.9.0
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
fiona==1.9.5
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
gcsfs==2024.2.0
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.17.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.3
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
Js2Py==0.74
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.2
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mercantile==1.2.1
mistune==3.0.2
mlflow==2.10.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.24.4
oauthlib==3.2.2
objsize==0.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
orjson==3.9.13
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
patsy==0.5.6
pendulum==3.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.4.0
prison==0.2.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
Protego==0.3.0
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.8
psycopg==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.0
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.17.2
pyjsparser==2.7.1
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyproj==3.6.1
pyrobuf==0.9.3
pyrosm==0.6.2
pytest==8.0.2
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-nvd3==0.15.0
python-rapidjson==1.14
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
querystring-parser==1.2.4
queuelib==1.6.2
rasterio==1.3.9
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.12.0
Scrapy==2.11.0
scrapyd==1.4.3
scrapyd-client==1.2.3
seaborn==0.13.2
Send2Trash==1.8.2
service-identity==23.1.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
statsmodels==0.14.1
streamlit==1.31.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
threadpoolctl==3.2.0
time-machine==2.13.0
tinycss2==1.2.1
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.38.1
Twisted==22.10.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
uberegg==0.1.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.0
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.27.0.post1
validators==0.22.0
w3lib==2.1.2
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.16.0
WTForms==3.1.2
xgboost==2.0.3
xxhash==3.4.1
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
zstandard==0.22.0
